Add files via upload
Okay github will not let me upload my zip file so im not sure if this file contains the changes I made. 
I added a nav bar that looks cleaner, we can add the logo as a nav bar brand but it wouldn't work for me right now.
Added a carousel but that also didn't run as planned but i added the pictures. Needs to be played around with. Again hopefully this is the new file.